feat: move header logic to the page
